+++ backend/tests/app/routers/test_qa.py
@@ -70,7 +70,8 @@ def test_chat_streamed(app_client, httpx_mock, patch_required_env, db_connection
 
     with app_client as app_client:
         create_output = app_client.post(
-            "/threads?virtual_lab_id=test_vlab&project_id=test_project"
+            "/threads",
+            json={"virtual_lab_id": "test_vlab", "project_id": "test_project"},
         ).json()
         response = app_client.post(
             f"/qa/chat_streamed/{create_output['thread_id']}",