+++ frontend/src/lib/cookies-server.ts
@@ -3,7 +3,7 @@ import { cookies } from "next/headers";
 export async function getSettings() {
   const cookieStore = await cookies();
   return {
-    projectID: cookieStore.get("projectID")?.value ?? "",
-    virtualLabID: cookieStore.get("virtualLabID")?.value ?? "",
+    projectID: cookieStore.get("projectID")?.value,
+    virtualLabID: cookieStore.get("virtualLabID")?.value,
   };
 }