+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Keywords filtering for literature search.
 - [frontend] - bugfixes.
 - Suggestion endpoint.
+- Support empty vlab+project in threads.
 
 ### Fixed
 - [frontend] - Parallel execution of tools
@@ -85,6 +86,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug for AI + tool calls.
 - [frontend] - Fix docker compose redirect bug.
 - [frontend] - Refactor
+- Websearch tools.
 
 ## [0.3.3] - 30.10.2024